client tweaks
